+++ R/to_brams_spm.R
@@ -1,7 +1,7 @@
-#' Inputs for BRAMS SPM
+#' Inputs for BRAMS-SPM
 #'
-#' @description Create inputs for SPM brams. The inputs consists in data-frame
-#' or a lists of data-frames with daily emissions (mol/day), lat, long. Also,
+#' @description Create inputs for BRAMS-SPM. The inputs consist of a data-frame
+#' or a list of data-frames with daily emissions (mol/day), lat, long. Also,
 #' including a functions describing the hourly profile.
 #'
 #' @param sdf Grid emissions, which can be a SpatialPolygonsDataFrame or polygon
@@ -12,8 +12,8 @@
 #' coordinates of the output.
 #' @return data-frame of daily gridded emissions, lat, long and a message with
 #' function.
-#' @note When the input os class 'Spatial', they are converted to 'sf'. If the
-#' input is a data-frame, the output is a data-frame. If he input is a list,
+#' @note When the input is class 'Spatial', they are converted to 'sf'. If the
+#' input is a data-frame, the output is a data-frame. If the input is a list,
 #' the output is a list.
 #'
 #' @references SPM BRAMS: FREITAS, E. MARTINS, L.,  SILVA, P. and ANDRADE, M.
@@ -21,46 +21,54 @@
 #' concentration forecast in the metropolitan area of são paulo, brazil:
 #' Coupling and validation. Atmospheric Environment, Elsevier, n. 39, p.
 #' 6352–6361, 2005.
-#' @author Sergio Ibarra
+#' @author Sergio Ibarra and Edmilson Freitas
 #' @importFrom sf st_as_sf st_coordinates st_set_geometry st_transform
 #' @export
 #'
 #' @examples \dontrun{
-#' # Do not run
-#'
+#' data(gCO)
+#' df1 <- to_brams_spm(sdf = gCO, epsg = 4326)
+#' head(df1)
+#' df2 <- to_brams_spm(sdf = list(co = gCO, pm = gCO), epsg = 4326)
+#' lapply(df2, head)
 #'}
 to_brams_spm <- function(sdf, epsg = 4326){
   if(inherits(x = sdf, what = "list")){
-    if(class(sdf[[1]]) == "SpatialPolygonsDataFrame"){
-      message("SpatialPolygonsDataFrame")
+    if(class(sdf[[1]])[1] == "SpatialPolygonsDataFrame"){
+      # message("SpatialPolygonsDataFrame")
       sdf <- lapply(sdf, sf::st_as_sf)
-    } else if(class(sdf[[1]] == "sf")){
-        message("sf")
-      }
-      dft <- as.data.frame(sf::st_coordinates(sf::st_transform(sdf[[1]], epsg)))
-      ldf <- lapply(1:length(sdf),
-                    function(i){
-                      cbind(colSums(sf::st_set_geometry(sdf[[i]], NULL)),
-                            dft)
-                    })
-      return(ldf)
-      # Initially, this function return rowsums and polygon separatly
-      sumdf <- sapply(sdf, rowSums, na.rm = T)
-      names(sumdf) <- paste0("sum_" , names(sdf))
-      print(sumdf)
-  } else if(class(sdf) == "sf"){
-    message("sf")
-    dft <- as.data.frame(sf::st_coordinates(sf::st_transform(sdf, epsg)))
-    ldf <- as.data.frame(cbind(colSums(sf::st_geometry(sdf, NULL)), dft))
+    } else if(class(sdf[[1]])[1] == "sf"){
+
+    }
+    dft <- as.data.frame(sf::st_coordinates(sf::st_transform(sdf[[1]], epsg)))
+    dft$N <- paste0(dft[, 3], "_", dft[, 4])
+    dft <- dft[!duplicated(dft$N),]
+    dft <- dft[, 1:2]
+    names(dft) <- c("long", "lat")
+    ldf <- lapply(1:length(sdf),
+                  function(i){
+                    cbind(rowSums(sf::st_set_geometry(sdf[[i]], NULL)),
+                          dft)
+                  })
+    nombres <- names(sdf)
+    for(i in 1:length(sdf)){
+      names(ldf[[i]]) <- c(nombres[i], "long", "lat")
+    }
     return(ldf)
-  } else if(class(sdf) == "SpatialPolygonsDataFrame") {
-    message("SpatialPolygonsDataFrame")
-    sdf <- sf::st_as_sf(sdf)
-    dft <- as.data.frame(sf::st_coordinates(sf::st_transform(sdf, epsg)))
-    ldf <- as.data.frame(cbind(colSums(sf::st_set_geometry(sdf, NULL)), dft))
-    sumdf <- rowSums(sdf, na.rm = T)
+    # Initially, this function return rowsums and polygon separatly
+    sumdf <- sapply(sdf, rowSums, na.rm = TRUE)
     names(sumdf) <- paste0("sum_" , names(sdf))
     print(sumdf)
+  } else {
+    # message("sf")
+    sdf <- sf::st_as_sf(sdf)
+    dft <- as.data.frame(sf::st_coordinates(sf::st_transform(sdf, epsg)))
+    dft$N <- paste0(dft[, 3], "_", dft[, 4])
+    dft <- dft[!duplicated(dft$N),]
+    dft <- dft[, 1:2]
+    names(dft) <- c("long", "lat")
+    dft2 <- data.frame(e24h = rowSums(sf::st_set_geometry(sdf, NULL)))
+    ldf <- cbind(dft, dft2)
     return(ldf)
   }
 }